Add products-csr to homepage and change url API
