+++ next.config.js
@@ -3,9 +3,9 @@ const nextConfig = {
   reactStrictMode: true,
   trailingSlash: true,
   images: {
-    domains: ['fakestoreapi.com'],
+    domains: ['naszsklep-api.vercel.app'],
     imageSizes: [16, 32, 48, 64, 96, 128, 256, 384],
   },
-}
+};
 
-module.exports = nextConfig
+module.exports = nextConfig;
